fix(prescriptions): corrected DTO mapping and controller structure
